requirements.txt
python-dotenv==1.0.0
requests==2.31.0
msgraph-core==0.2.2
azure-identity==1.14.0
msal==1.24.0
pydantic==2.10.0
pyyaml==6.0.3
pypdf>=4.0.0
pdfplumber==0.10.3
openpyxl==3.1.5
python-dateutil==2.8.2
Pillow==10.4.0
click==8.1.7
tabulate==0.9.0
openai==1.55.3
beautifulsoup4==4.12.3
pyodbc>=5.0
